moves warnings to wt project and adds yellow color
